+++ src/almanack/reporting/report.py
@@ -2,42 +2,8 @@
 This module creates entropy reports
 """
 
-import json
-import pathlib
 from typing import Dict, Any 
-import tabulate
-
-from almanack.processing.processing_repositories import process_entire_repo
-
-
-def process_repo_entropy(repo_path: str) -> None:
-    """
-    CLI entry point to process a repository for calculating entropy changes between commits and
-    generate a report.
-
-    Args:
-        repo_path (str): The local path to the Git repository.
-    """
-
-    repo_path = pathlib.Path(repo_path)
-
-    # Check if the directory contains a Git repository
-    if not repo_path.exists() or not (repo_path / ".git").exists():
-        raise FileNotFoundError(f"The directory {repo_path} is not a repository")
-
-    # Process the repository and get the dictionary
-    entropy_data = process_entire_repo(str(repo_path))
-
-    # Generate and print the report from the dictionary
-    report_content = whole_repo_report(entropy_data)
-
-    # Convert the dictionary to a JSON string
-    json_string = json.dumps(entropy_data, indent=4)
-
-    print(report_content)
-
-    # Return the JSON string and report content
-    return json_string
+from tabulate import tabulate
 
 
 def whole_repo_report(data: Dict[str, Any]) -> str:
@@ -60,11 +26,11 @@ def whole_repo_report(data: Dict[str, Any]) -> str:
     time_range_of_commits = data["time_range_of_commits"]
     entropy_data = data[
         "file_level_entropy"
-    ]  # Assume this contains entropy for each file
+    ] 
 
     # Sort files by normalized entropy in descending order and get the top 5
     sorted_entropy = sorted(
-        entropy_data.items(), key=operator.itemgetter(1), reverse=True
+        entropy_data.items(), key=lambda item:item[1], reverse=True
     )
     top_files = sorted_entropy[:5]
