+++ tests/test_generate_data.py
@@ -6,7 +6,7 @@ import pathlib
 import shutil
 import tempfile
 
-from almanack.processing.processing_repositories import process_entire_repo
+from almanack.processing.generate_data import generate_whole_repo_data
 
 
 def test_process_entire_repo(repository_paths: dict[str, pathlib.Path]) -> None:
@@ -22,7 +22,7 @@ def test_process_entire_repo(repository_paths: dict[str, pathlib.Path]) -> None:
             shutil.copytree(repo_path, temp_repo_path)
 
             # Call the function with the temporary repository path
-            entropy_data = process_entire_repo(str(temp_repo_path))
+            entropy_data = generate_whole_repo_data(str(temp_repo_path))
 
             # Check if the entropy data is not empty
             assert entropy_data
